requirements.txt
langchain 
llama-index 
llama_hub
accelerate
pyngrok
chromadb
pandas 
faiss-cpu 
sentence-transformers 
transformers 
langchain-community
fastapi 
uvicorn
protobuf